+++ src/main/resources/markdown-macro-config.vm
@@ -29,13 +29,20 @@
 	    <div class="aui-message aui-message-confirmation" id="save-success-popup" style="display: none;">
 	        <p>Changes saved successfully.</p>
 	    </div>
+        
+        <h3 style="padding: 10px 0;">Enable Azure DevOps support.</h3>
 
-        <h3 style="padding: 10px 0;">Enable Azure DevOps custom relative path resolution. Otherwise classic github-style relative path support is used.</h3>
-
+        <div class="flex-box" style="margin-bottom: 10px;">
+            <ul>
+                <li>Relative paths can be configured to be resolved to Azure DevOps repository.</li>
+                <li>Inline image resize syntax. Eg: ![](/src =WxH)</li>
+            </ul>
+        </div>
+        
         <div class="flex-box">
             <div class="checkbox">
                 <input type="checkbox" id="azure-customs-enabled" class="checkbox" name="azure-customs-enabled">
-                <label for="azure-customs-enabled">Enable Azure DevOps Relative Paths for Markdown From Url Macro</label>
+                <label for="azure-customs-enabled">Enable Azure DevOps support</label>
             </div>
         </div>
